[Libra Node] Remove main_node module and merge into lib
